+++ src/Provider/UserId/AuthenticationServiceFactory.php
@@ -1,31 +1,41 @@
 <?php
 /**
  * @license   http://opensource.org/licenses/BSD-3-Clause BSD-3-Clause
- * @copyright Copyright (c) 2015 Zend Technologies USA Inc. (http://www.zend.com)
+ * @copyright Copyright (c) 2015-2016 Zend Technologies USA Inc. (http://www.zend.com)
  */
 
 namespace ZF\OAuth2\Provider\UserId;
 
-use Zend\ServiceManager\FactoryInterface;
-use Zend\ServiceManager\ServiceLocatorInterface;
+use Interop\Container\ContainerInterface;
 
-class AuthenticationServiceFactory implements FactoryInterface
+class AuthenticationServiceFactory
 {
     /**
-     * @param ServiceLocatorInterface $services
+     * @param  ContainerInterface $container
      * @return AuthenticationService
      */
-    public function createService(ServiceLocatorInterface $services)
+    public function __invoke(ContainerInterface $container)
     {
-        $config = $services->get('Config');
+        $config = $container->get('config');
 
-        if ($services->has('Zend\Authentication\AuthenticationService')) {
+        if ($container->has('Zend\Authentication\AuthenticationService')) {
             return new AuthenticationService(
-                $services->get('Zend\Authentication\AuthenticationService'),
+                $container->get('Zend\Authentication\AuthenticationService'),
                 $config
             );
         }
 
         return new AuthenticationService(null, $config);
+    }
+
+    /**
+     * Provided for backwards compatibility; proxies to __invoke().
+     *
+     * @param \Zend\ServiceManager\ServiceLocatorInterface $container
+     * @return AuthenticationService
+     */
+    public function createService($container)
+    {
+        return $this($container);
     }
 }